apps/api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.6
python-multipart==0.0.20
boto3==1.35.91
psycopg[binary]==3.2.4
SQLAlchemy==2.0.36
alembic==1.14.0
langchain-openai==0.2.14
python-dotenv==1.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
